refactor: Button 컴포넌트 프롭스로 구분 -> 합성컴포넌트 방식으로 변경
